deepfake-main/deepfake_detection/backend/deepfake-master/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-jwt-extended>=4.4.0
werkzeug>=2.3.0
numpy>=1.24.0
Pillow>=10.0.0
bcrypt>=4.0.0
